Small Refactor (move all related widgets to subfolder)
