+++ launcher/minecraft/gameoptions/GameOptionDelegate.h
@@ -37,6 +37,7 @@ class GameOptionDelegate : public QStyledItemDelegate {
                                                    const QStyleOptionViewItem& option,
                                                    const QModelIndex& index) const override;
     void setModelData(QWidget* editor, QAbstractItemModel* model, const QModelIndex& index) const override;
+    QSize sizeHint(const QStyleOptionViewItem& option, const QModelIndex& index) const override;
 
     enum GameOptionWidgetType { slider, text, keybind, number, comboBox };
