+++ launcher/ui/widgets/GameOptions/GameOptionWidgetCheckBox.cpp
