+++ launcher/ui/widgets/GameOptions/GameOptionWidgetCheckBox.h
