+++ launcher/ui/widgets/GameOptions/GameOptionWidgetCheckBox.ui
