+++ launcher/ui/widgets/GameOptions/GameOptionWidgetComboBox.cpp
