+++ launcher/ui/widgets/GameOptions/GameOptionWidgetSlider.ui
