+++ launcher/ui/widgets/GameOptions/GameOptionWidgetSpinnerFloat.cpp
