+++ launcher/ui/widgets/GameOptions/GameOptionWidgetSpinnerFloat.h
