+++ launcher/ui/widgets/GameOptions/GameOptionWidgetSpinnerFloat.ui
