+++ launcher/ui/widgets/GameOptions/GameOptionWidgetSpinnerInt.cpp
