+++ launcher/ui/widgets/GameOptions/GameOptionWidgetSpinnerInt.h
