+++ launcher/ui/widgets/GameOptions/GameOptionWidgetSpinnerInt.ui
